🐛 Fix bug breaking empty buckets functionality if the bucket is empty

+++ pkg/s3api/s3api.go
@@ -105,6 +105,10 @@ func (a *S3API) deleteAllObjects(bucketName string) error {
 			return err
 		}
 
+		if result.Contents == nil {
+			break
+		}
+
 		_, err = a.provider.S3().DeleteObjects(&s3.DeleteObjectsInput{
 			Bucket: aws.String(bucketName),
 			Delete: &s3.Delete{Objects: objectsAsIdentifiers(result.Contents)},
